requirements.txt
fastapi==0.63.0
uvicorn==0.13.3
PyJWT==2.0.1
requests==2.25.1
SQLAlchemy==1.3.22
python-dotenv==0.15.0
PyMySQL==0.10.1
